+++ e_ood/__init__.py
@@ -4,5 +4,5 @@ __version__ = '0.1.40a1'
 
 from .analyze import Analyzer  # noqa
 from .db import ReportedUpdateTypes, VersionDB  # noqa
-from .pypi import PackageVersionInfo  # noqa
+from .pypi import AvailablePackageVersions  # noqa
 from .virtualenv import EnvPackages  # noqa